added ZeRO optim

+++ src/training/losses.py
@@ -87,7 +87,7 @@ class SimCTGLoss( nn.Module ):
         loss_matrix = torch.nn.functional.relu(loss_matrix)
 
         ### input mask
-        input_mask = torch.ones_like( input_ids ).type( torch.FloatTensor ) # type: ignore
+        input_mask = torch.ones_like( input_ids ).type( torch.float16 ) # type: ignore
         if loss_matrix.is_cuda:
             input_mask = input_mask.cuda(loss_matrix.get_device())
         input_mask = input_mask.masked_fill(input_ids.eq(self.pad_token_id), 0.0)
